perf(output): Replace Handlebars with direct string building for markdown and plain output (-2.3% to -2.6%)
Replace Handlebars template execution with direct Array.push() + join()
string building for markdown and plain output formats, matching the
approach already used for XML output (commit d72f102).

Why: Handlebars adds ~93ms overhead for non-XML formats: ~27ms for the
dynamic import, ~5ms for template compilation, and ~15ms for template
execution with per-iteration function call dispatch. The remaining gap
is hidden by pipeline parallelism with metrics workers.

What changed:
- Add generateDirectMarkdownOutput() using Array.push()+join() pattern
- Add generateDirectPlainOutput() using the same approach
- Extract extensionToLanguageMap + getLanguageFromFilePath() into new
  outputLanguageMap.ts to avoid pulling Handlebars into the direct path
- Route non-parsable markdown/plain through direct builders in
  generateOutput(), skipping createRenderContext entirely
- Update test mocks to include the two new deps functions

Output is byte-for-byte identical to the Handlebars templates.

Benchmark (20 alternating A/B measurements, paired t-test):
  --style markdown: -26.8ms mean (-2.6%), paired t=3.89 (p<0.001)
  --style plain:    -24.1ms mean (-2.3%), paired t=3.46 (p<0.005)

Handlebars is still lazy-loaded for parsable markdown/plain output
(--parsable-style flag), which goes through createRenderContext.
